Open Store Filter
Adding of the "OpenStore" component, which was created as a switch to
enable the user to decide wheter he wants to filter the search by the
opened stores or not.

Changes in several files that were filtering information by the opened
stores, in order to improve performance and code readability.

Changes in the way the properties "firebase_id_store",
"shipping_tax", "opening_time" and "closing_time" are passed to the
products in the file "Search/index.js", in order to improve code
readability.